Get person by id query has been added

+++ src/ArtExchange.Application/Contracts/Repository/IRepositoryAsync.cs
@@ -4,7 +4,7 @@ namespace ArtExchange.Application.Contracts.Repository
 {
     public interface IRepositoryAsync<T> where T : class
     {
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
         Task<IReadOnlyList<T>> ListAllAsync();
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
